Add VLC audio backend
